# [FIX] purchase_stock: compute_picking perfs improvements