Avoid calling two mapped in nested for loop and use a defaultdict
instead to improve compute_picking performances.

#### speedup

Client DB, 7550 purchase_order, `order._compute_picking` average time. Preprocessed orders
by calling `_create_picking` until `_action_assign` to create stock.moves beforehand

| # Pickings | Before PR | After PR |
|:----------:|:-----------:|:--------:|
| 10 | 0.854s | 0.229s |
| 50 | 1.406s | 0.498s |
| 200 | 3.664s | 1.5s |
| 1000 | 30.19s | 10.33s |
| 2500 | 58.64s | 22.04s |
| 7550 | 2m15s | 57.45s |

`order._compute_picking` average time by # stock_moves for 1 picking.

| # Stock Moves | Before PR | After PR |
|:---------------:|:----------:|:---------:|
| 10 | 0.053s | 0.027s |
| 25 | 0.075s | 0.031s |
| 50 | 0.115s | 0.039s |
| 100 | 0.192s | 0.055s |
| 246 | 0.418s | 0.099s |


--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr