@Aurelienvd 2021-06-22
Hi @ryv-odoo I checked v13 and 14 before committing but I wouldn't have thought changes like that were made in master. I backported PR odoo/odoo#35878, only adding an explicit `mapped` because `Relational.__get__` seems to have been introduced in v13.

From my tests it seems to be working in 12.0 and the performances are fine, the speedup is roughly the same, maybe slightly faster than the one reported above.